Split the brutus-api views.py module into a package with sub-modules.
