+++ src/brutus-api/brutus_api/views/__init__.py
@@ -1,0 +1,3 @@
+from .general import index
+from .modules import modules
+from .requests import requests, get_request
