+++ src/brutus-api/brutus_api/views/general.py
@@ -1,0 +1,10 @@
+from brutus_api.app import app
+
+
+@app.route('/')
+def index():
+    """
+    Get the index page.
+    """
+
+    return "Brutus API"
